refs #48 Refactor status actions
